requirements.txt
Flask
gunicorn
pandas
werkzeug
redis>=5.0
azure-core>=1.29.0
azure-cosmos>=4.5.0
scikit-learn
numpy